refactored episodecollector to storage
